IGAPP-652: Move news components
