+++ web-ts/src/components/NewsListItem.tsx
@@ -8,10 +8,10 @@ import LastUpdateInfo from '../../../modules/common/components/LastUpdateInfo'
 import { Description } from '../../../modules/common/components/ListItem'
 import { type TFunction } from 'react-i18next'
 import textTruncator from '../../../modules/common/utils/textTruncator'
-import { LOCAL_NEWS } from '../constants'
-import type { NewsType } from '../constants'
-import DateFormatter from 'api-client/src/i18n/DateFormatter'
-import type { ThemeType } from 'build-configs/ThemeType'
+import { LOCAL_NEWS } from '../../../web/src/routes/news/constants'
+import type { NewsType } from '../../../web/src/routes/news/constants'
+import DateFormatter from '../api-client/src/i18n/DateFormatter'
+import type { ThemeType } from '../build-configs/ThemeType'
 import { Parser } from 'htmlparser2'
 
 export const NUM_OF_WORDS_ALLOWED = 30
