+++ web-ts/src/components/NewsTabs.tsx
@@ -4,10 +4,10 @@ import * as React from 'react'
 import styled, { type StyledComponent } from 'styled-components'
 import { type TFunction } from 'react-i18next'
 import Tab from './Tab'
-import { LOCAL_NEWS, TU_NEWS } from '../constants'
-import LocalNewsRouteConfig from '../../../modules/app/route-configs/LocalNewsRouteConfig'
-import TunewsRouteConfig from '../../../modules/app/route-configs/TunewsRouteConfig'
-import type { ThemeType } from 'build-configs/ThemeType'
+import { LOCAL_NEWS, TU_NEWS } from '../../../web/src/routes/news/constants'
+import LocalNewsRouteConfig from '../../../web/src/modules/app/route-configs/LocalNewsRouteConfig'
+import TunewsRouteConfig from '../../../web/src/modules/app/route-configs/TunewsRouteConfig'
+import type { ThemeType } from '../build-configs/ThemeType'
 
 const StyledTabs: StyledComponent<{||}, ThemeType, *> = styled.div`
   display: flex;
