--- web/src/routes/news/constants.js
@@ -1,6 +1,0 @@
-// @flow
-
-export type NewsType = 'LOCAL_NEWS' | 'TU_NEWS'
-
-export const LOCAL_NEWS = 'LOCAL_NEWS'
-export const TU_NEWS = 'TU_NEWS'
